Add defects "template" and URL pattern.
The defects page takes a URL of the pattern /defects/<textid>/ and
does a search for all the words in the specified text with the
preselected flag set, which we typically use to identify gaps and
other low-hanging fruit.

This isn't really a template but just a static page with a form
that mimics the search form but sets the text based on the URL.
It violates DRY in that if the search form on the main page is
ever changed, this will have to be changed to match, but it's
just 25 lines of very simple HTML and doesn't require touching
the Django code except for adding one line to urls.py.