Add a library finder button to the search form.
It's not really part of the search form but is its own form. The
button will open a new window to the WUSTL library finder site. If
no text is selected in AnnoLex, it will open a list of all libraries
in the library finder database, but if a text is selected in AnnoLex,
the library finder will display only libraries containing copies of
that text.